Extract a shared TrackedPath base for asset and user tracks
SMMAsset and SMMUserPosition were near-identical: accumulated path,
track colour with its persisted cookie/swatch, the colour-picker dialog,
and the in-flight-guarded incremental history fetch were duplicated, the
only differences being the colour cookie key and the history endpoint.

Pull the shared behaviour into TrackedPath (components/TrackedPath.tsx)
with two abstract hooks - colorCookieKey() and historyUrl() - and have
both classes extend it. SMMUserPosition keeps its own popup field. Pure
refactor; behaviour is unchanged (the `from` guard is normalised to the
already-equivalent `!= null`).
